PSPDFKit 2.14.0 - Adds support to Copy/Paste annotations, FreeText annotation editing is now always sharp, the new global style manager remembers last used annotation settings, performance improvements for large documents, improved tinting feature, several smaller tweaks and bug fixes.
Former-commit-id: 4a10dd94b83377fb31409283294ffaf0a91828ea [formerly fcc79bedccb90a148dc271a600c806ee86a7e704]
Former-commit-id: 50d12258edd199aa1f718ed1a1f60a16b2c7b66f

+++ PSPDFKit.embeddedframework/PSPDFKit.framework/Versions/A/Headers/PSPDFAction.h
@@ -7,7 +7,7 @@
 
 #import "PSPDFModel.h"
 
-typedef NS_ENUM(NSUInteger, PSPDFActionType) {
+typedef NS_ENUM(UInt8, PSPDFActionType) {
     PSPDFActionTypeURL,
     PSPDFActionTypeGoTo,
     PSPDFActionTypeRemoteGoTo,
